# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de veículos, revisões e manutenções do CVMC (Como Vai Meu Carro).",
        "title": "CVMC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte CVMC",
            "email": "suporte@cvmc.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/cars": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos os veículos pertencentes ou compartilhados com o usuário autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Listar veículos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo veículo associado à conta do usuário autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Cadastrar novo veículo",
                "parameters": [
                    {
                        "description": "Dados do veículo",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/cars/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados completos de um veículo específico pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Obter detalhes do veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados cadastrais e quilometragem do veículo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Atualizar veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove o veículo do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Excluir veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/cars/{id}/share": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Concede permissão de acesso ao veículo para outro usuário através do e-mail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Compartilhar veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "E-mail do usuário",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ShareCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/cars/{id}/share/{userID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoga o acesso de um usuário ao veículo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Remover compartilhamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do usuário a desvincular",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateCarRequest": {
            "type": "object",
            "properties": {
                "lastMileage": {
                    "type": "integer",
                    "example": 32000
                },
                "manufacturer": {
                    "type": "string",
                    "example": "Honda"
                },
                "model": {
                    "type": "string",
                    "example": "Civic Touring"
                },
                "name": {
                    "type": "string",
                    "example": "Meu Civic"
                },
                "yearManufacture": {
                    "type": "integer",
                    "example": 2023
                },
                "yearModel": {
                    "type": "integer",
                    "example": 2024
                }
            }
        },
...
        "handlers.ShareCarRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "co-motorista@cvmc.com"
                }
            }
        },
        "handlers.UpdateCarRequest": {
            "type": "object",
            "properties": {
                "lastMileage": {
                    "type": "integer",
                    "example": 35000
                },
                "manufacturer": {
                    "type": "string",
                    "example": "Honda"
                },
                "model": {
                    "type": "string",
                    "example": "Civic Touring"
                },
                "name": {
                    "type": "string",
                    "example": "Meu Civic"
                },
                "yearManufacture": {
                    "type": "integer",
                    "example": 2023
                },
                "yearModel": {
                    "type": "integer",
                    "example": 2024
                }
            }
        },
...
        "httpx.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "httpx.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}